requirements.txt
streamlit==1.28.2
Pillow==10.1.0
numpy==1.26.2 